feat: enhance diet management by adding user association, updating API routes, and implementing user registration and login functionalities

+++ app/api/diet/deleteDiet.ts
@@ -1,8 +1,8 @@
-import api from "./axios"
+import api from "../axios"
 
-const deleteDiet = async (id: string) => {
+const deleteDiet = async (userId: string, id: string) => {
   try {
-    const response = await api.delete(`/diets/${id}`)
+    const response = await api.delete(`/diets/${userId}/${id}`)
     return response.data
   } catch (error) {
     console.error('Erro ao deletar dieta:', error)
